+++ R/sNNclust.R
@@ -83,7 +83,7 @@
 #' cl <- sNNclust(DS3, k = 20, eps = 7, minPts = 16)
 #' plot(DS3, col = cl$cluster + 1L, cex = .5)
 #'
-#' @export sNNclust
+#' @export
 sNNclust <- function(x, k, eps, minPts, borderPoints = TRUE, ...) {
   nn <- sNN(x, k=k, jp = TRUE, ...)